fix bug where input byte would be skipped when restarting pattern
The pattern_wild_full() testcase (as of last commit) demonstrated this
failure. The problem was that we would skip checking an input byte when
we started the pattern over. b++ would happen unconditionally, even when
p was reset to zero, causing us to not re-check that input byte when p=0.