mmap() input file if possible
If HAS_MMAP is defined, mmap(2) is used to map the file into memory,
instead of malloc + fread. This slightly improves performance, and
drastically reduces memory usage.

This effectively removes the possibility of reading from stdin. Because
a pattern could, in theory, match the entire input file, we must have
the entire file in memory. This removal makes supporting mmap easier
architecturally.